qcanpool: add FancyTabBar and FancyTabWidget
Both tab and action use QToolButton.

When there is a menu in QToolButton and it is displayed vertically,
the icon display will not be vertically aligned with other icons.
This requires customizing special controls based on QToolButton.

In fancydemo, I found that when the main window is resized,
mouse events are easily captured by the upper-level controls,
resulting in abnormal mouse state or invalid resizing.
FancyTitleBar or FancyBar should be optimized.

Signed-off-by: canpool <canpool@163.com>