qxribbon: RibbonQuickAccessBar: perfect control action visible
If a QAction is added to both RibbonQuickAccessBar and RibbonGroup,
when RibbonQuickAccessBar hides the QAction through setVisible (false),
and the QAction in RibbonGroup will also be hidden.

Therefore, the RibbonQuickAccessBar achieves the show and hiding of
QAction through addAction and removeAction (simulating takeAction).

However, QAction can also be hidden alone or within the RibbonGroup
through setVisible(false). Therefore, QAction is also hidden in the
RibbonQuickAccessBar. That is to say, RibbonQuickAccessBar does not
need to hide QAction through removeAction.

In summary, RibbonQuickAccessBar combines setVisible,
addAction/insertAction, and removeAction (simulating takeAction) to
achieve the show and hiding of QAction.

Signed-off-by: canpool <canpool@163.com>