qxribbon: qss: set background for MenuButtonPopup button
The purpose of setting the background color is to enable a clear
observation of the buttons and the menu area when the mouse hovers over
them.

The appearance of the MenuButtonPopup button in QSS should be
represented as xxxButton[popupMode="1"]. If the background color of the
button is set using this form, many configurations need to be added.
Therefore, for simplicity, the background color has been directly set
for the button here.

Signed-off-by: canpool <canpool@163.com>