qxribbon: RibbonWindow: fix the shown when setting frameless
If you directly show the main window when setting frameless, the
following problem will occur: If setFrameless is called before the main
window is shown, it will cause the main window to first show an empty
small window and then the formal window.

Therefore, only in the case where the window changes from shown to
hidden is the show() function actively called.

Signed-off-by: canpool <canpool@163.com>