qxribbon: RibbonMainWindow: minimum width also needs to be set in Resize event
When the main window uses the system native frame, when the main window
is restored from the maximum for the first time, the restored width is
smaller than the minimum width of the RibbonBar, but the Resize event of
the RibbonBar will not be sent. Therefore, the minimum width also needs
to be set here.

Signed-off-by: canpool <canpool@163.com>